+++ apps/sugarfoot/src/events/events.module.ts
