+++ apps/tasks/src/consumer/project-events.interface.ts
@@ -1,4 +1,4 @@
 export interface ProjectCreatedEvent {
   id: string;
-  userId: string;
+  ownerId: string;
 }
